debian/rules: Adapt to new upstream build system.
Upstream now uses a method of building a monolithic executable with a python
shebang line with the code appended in a zip file.

To work around this and have a regular Python application, I use make to
generate the files (like documentation etc.), move the Makefile aside (since
that's what dh_auto_build would use) and let the regular Python setup.py be
used to install everything.